+++ packages/sdk/src/sdk/notesManager/mock.ts
@@ -1,14 +1,14 @@
 /* eslint-disable */
 import { NotesManager, JoinSplitEvent } from ".";
 import { NocturnePrivKey } from "../../crypto";
-import { LocalObjectDB } from "../db/local";
+import { NotesDB, InMemoryKVStore } from "../db";
 import { IncludedNote } from "../note";
 import { NocturneSigner } from "../signer";
 
 export class MockNotesManager extends NotesManager {
   constructor() {
     super(
-      new LocalObjectDB({ localMerkle: true }),
+      new NotesDB(new InMemoryKVStore()),
       new NocturneSigner(new NocturnePrivKey(1n))
     );
   }